Added new separate class for handling attachements
